Use 128bit fileId seed rather than 64bit (#167)

+++ backend/canisters/bucket/impl/src/model/files.rs
@@ -172,7 +172,7 @@ impl Files {
         caller: UserId,
         file_id: FileId,
         canister_id: CanisterId,
-        file_id_seed: u64,
+        file_id_seed: u128,
         accessors: HashSet<UserId>,
         now: TimestampMillis,
     ) -> ForwardFileResult {
